Filter channel list case-insensitively
Mobile browsers default to a "shifted" keyboard when you focus a text field. With case-sensitive filtering you have to remember to un-shift, or the suggestion list will always be empty. 

There's also an `autocapitalize="none"` attribute that would mostly fix the problem, but I don't think it's supported in mobile Firefox.