Disable the Upload button when it's clicked
On mobile in particular, it can be hard to tell if you successfully
clicked the button or clumsily scrolled the page. The share-to-slack
function can take several seconds and there's no indication that it's in
progress, which exacerbates the problem.

Setting the button to disabled gives a visual indication that the share
is indeed in progress.